# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pelayanan Scheduler API",
    "description": "API for managing church service scheduling including users, positions, categories, and worship services",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": ""
      }
    },
    "schemas": {
...
      "KRW": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "krw_name": { "type": "string", "example": "KRW Jakarta" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/krw": {
      "get": {
        "summary": "Get all KRWs",
        "description": "Retrieve all KRW entries",
        "tags": ["KRW"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "required": false,
            "description": "Search KRW by name"
          },
          {
            "in": "query",
            "name": "skip",
            "schema": { "type": "integer", "default": 0 },
            "required": false,
            "description": "Number of records to skip (for pagination)"
          },
          {
            "in": "query",
            "name": "take",
            "schema": { "type": "integer", "default": 10, "maximum": 100 },
            "required": false,
            "description": "Number of records to return (max 100)"
          }
        ],
        "responses": {
          "200": {
            "description": "List of KRWs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/KRW" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new KRW",
        "description": "Create a new KRW entry",
        "tags": ["KRW"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KRW"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "KRW created successfully"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/api/krw/{id}": {
      "get": {
        "summary": "Get KRW by ID",
        "tags": ["KRW"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "KRW detail" },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "summary": "Update KRW",
        "tags": ["KRW"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KRW"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "KRW updated" },
          "400": { "description": "Bad Request" }
        }
      },
      "delete": {
        "summary": "Delete KRW",
        "tags": ["KRW"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" },
          "404": { "description": "Not Found" }
        }
      }
    },
...
    "/api/pelayan-level": {
      "get": {
        "summary": "Get all Pelayan Levels",
        "description": "Retrieve all level",
        "tags": ["Pelayan Levels"],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "required": false,
            "description": "Search Pelayanan Level by name"
          },
          {
            "in": "query",
            "name": "skip",
            "schema": { "type": "integer", "default": 0 },
            "required": false,
            "description": "Number of records to skip (for pagination)"
          },
          {
            "in": "query",
            "name": "take",
            "schema": { "type": "integer", "default": 10, "maximum": 100 },
            "required": false,
            "description": "Number of records to return (max 100)"
          }
        ],
        "responses": {
          "200": { "description": "List of positions" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/api/pelayan-level/{id}": {
      "get": {
        "summary": "Get Level by ID",
        "tags": ["Pelayan Levels"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Level detail" },
          "404": { "description": "Not Found" }
        }
      },
      "patch": {
        "summary": "Update Level",
        "tags": ["Pelayan Levels"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Level"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Level updated" },
          "400": { "description": "Bad Request" }
        }
      },
      "delete": {
        "summary": "Delete Level",
        "tags": ["Pelayan Levels"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "204": { "description": "Deleted successfully" },
          "404": { "description": "Not Found" }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}